Add names to tests
These names reflect those used in the reference implementation. This makes it
easier to reason about what tests are missing or wrong or need updating.

Note that there are some duplicate or wrong names in these; these reflect
the status quo in the reference implemenation. These are tracked in
paragonie/paseto#85.

Signed-off-by: Martijn Pieters <mj@zopatista.com>